+++ views/admin/index.php
@@ -1,8 +1,8 @@
 <?php
-/* @var $this \humhub\modules\ui\view\components\View */
+/* @var $this \humhub\components\View */
 /* @var $model \humhubContrib\auth\microsoft\models\ConfigureForm */
 
-use yii\bootstrap\ActiveForm;
+use humhub\widgets\form\ActiveForm;
 use yii\helpers\Html;
 
 ?>
@@ -13,7 +13,7 @@ use yii\helpers\Html;
 
         <div class="panel-body">
             <p>
-                <?= Html::a(Yii::t('AuthMicrosoftModule.base', 'Microsoft Documentation'), 'https://docs.microsoft.com/en-us/azure/active-directory/develop/active-directory-v2-protocols#app-registration', ['class' => 'btn btn-primary pull-right btn-sm', 'target' => '_blank']); ?>
+                <?= Html::a(Yii::t('AuthMicrosoftModule.base', 'Microsoft Documentation'), 'https://docs.microsoft.com/en-us/azure/active-directory/develop/active-directory-v2-protocols#app-registration', ['class' => 'btn btn-primary float-end btn-sm', 'target' => '_blank']); ?>
                 <?= Yii::t('AuthMicrosoftModule.base', 'Please follow the instructions to create the required <strong>client</strong> credentials.'); ?>
                 <br/>
             </p>
@@ -31,7 +31,7 @@ use yii\helpers\Html;
             <?= $form->field($model, 'redirectUri')->textInput(['readonly' => true]); ?>
             <br/>
 
-            <div class="form-group">
+            <div class="mb-3">
                 <?= Html::submitButton(Yii::t('base', 'Save'), ['class' => 'btn btn-primary', 'data-ui-loader' => '']) ?>
             </div>