+++ docs/CHANGELOG.md
@@ -8,3 +8,5 @@ Changelog
 1.0.0 (Unreleased)
 ------------------
 - Initial release
+- Enh #6: Use PHP CS Fixer
+- Enh #7: Migration to Bootstrap 5 for HumHub 1.18